ai_daily_briefing/requirements.txt
arxiv==2.1.3
google-generativeai==0.8.3
anthropic==0.49.0
requests>=2.32.0
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.0
lxml>=5.3.0
